+++ ElvUI_LuckyoneUI/Core/Commands.lua
@@ -24,10 +24,17 @@ function L1UI:Toggles(msg)
 	if msg == 'install' then
 		E.PluginInstaller:Queue(L1UI.InstallerData)
 	elseif msg == 'config' then
-		E:ToggleOptionsUI()
+		E:ToggleOptions()
 		E.Libs.AceConfigDialog:SelectGroup('ElvUI', 'L1UI')
 	end
 end
+
+local AddOns = {
+	ElvUI = true,
+	ElvUI_Libraries = true,
+	ElvUI_LuckyoneUI = true,
+	ElvUI_Options = true
+}
 
 -- LuckyoneUI debug mode
 function L1UI:DebugMode(msg)
@@ -35,7 +42,7 @@ function L1UI:DebugMode(msg)
 	if switch == 'on' then
 		for i = 1, GetNumAddOns() do
 			local name = GetAddOnInfo(i)
-			if name ~= 'ElvUI' and name ~= 'ElvUI_OptionsUI' and name ~= 'ElvUI_LuckyoneUI' and E:IsAddOnEnabled(name) then
+			if not AddOns[name] and E:IsAddOnEnabled(name) then
 				DisableAddOn(name, E.myname)
 				ElvDB.LuckyoneDisabledAddOns[name] = i
 			end